# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the freelance design client portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProjectProjectType =
  (typeof ProjectProjectType)[keyof typeof ProjectProjectType];

export const ProjectProjectType = {
  logo: "logo",
  branding: "branding",
  web_design: "web_design",
  social_media: "social_media",
  print: "print",
  illustration: "illustration",
  other: "other",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  pending: "pending",
  in_progress: "in_progress",
  review: "review",
  delivered: "delivered",
  completed: "completed",
} as const;

export type ProjectPriority =
  (typeof ProjectPriority)[keyof typeof ProjectPriority];

export const ProjectPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  projectType: ProjectProjectType;
  status: ProjectStatus;
  priority: ProjectPriority;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  budget: number | null;
  /** @nullable */
  deadline: string | null;
  createdAt: string;
  updatedAt: string;
  fileCount: number;
  messageCount: number;
}

export type ProjectFileUploadedBy =
  (typeof ProjectFileUploadedBy)[keyof typeof ProjectFileUploadedBy];

export const ProjectFileUploadedBy = {
  client: "client",
  designer: "designer",
} as const;

export interface ProjectFile {
  id: number;
  projectId: number;
  fileName: string;
  fileSize: number;
  fileType: string;
  uploadedBy: ProjectFileUploadedBy;
  uploadedAt: string;
  url: string;
}

export type MessageSenderRole =
  (typeof MessageSenderRole)[keyof typeof MessageSenderRole];

export const MessageSenderRole = {
  client: "client",
  designer: "designer",
} as const;

export interface Message {
  id: number;
  projectId: number;
  content: string;
  senderName: string;
  senderRole: MessageSenderRole;
  createdAt: string;
}

export type ProjectDetail = Project & {
  files: ProjectFile[];
  messages: Message[];
};

export type CreateProjectBodyProjectType =
  (typeof CreateProjectBodyProjectType)[keyof typeof CreateProjectBodyProjectType];

export const CreateProjectBodyProjectType = {
  logo: "logo",
  branding: "branding",
  web_design: "web_design",
  social_media: "social_media",
  print: "print",
  illustration: "illustration",
  other: "other",
} as const;

export type CreateProjectBodyPriority =
  (typeof CreateProjectBodyPriority)[keyof typeof CreateProjectBodyPriority];

export const CreateProjectBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateProjectBody {
  title: string;
  description: string;
  projectType: CreateProjectBodyProjectType;
  priority: CreateProjectBodyPriority;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  deadline?: string | null;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  review: "review",
  delivered: "delivered",
  completed: "completed",
} as const;

export type UpdateProjectBodyPriority =
  (typeof UpdateProjectBodyPriority)[keyof typeof UpdateProjectBodyPriority];

export const UpdateProjectBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  status?: UpdateProjectBodyStatus;
  priority?: UpdateProjectBodyPriority;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  deadline?: string | null;
}

export type UploadFileBodyUploadedBy =
  (typeof UploadFileBodyUploadedBy)[keyof typeof UploadFileBodyUploadedBy];

export const UploadFileBodyUploadedBy = {
  client: "client",
  designer: "designer",
} as const;

export interface UploadFileBody {
  fileName: string;
  fileSize: number;
  fileType: string;
  uploadedBy: UploadFileBodyUploadedBy;
  url: string;
}

export type CreateMessageBodySenderRole =
  (typeof CreateMessageBodySenderRole)[keyof typeof CreateMessageBodySenderRole];

export const CreateMessageBodySenderRole = {
  client: "client",
  designer: "designer",
} as const;

export interface CreateMessageBody {
  content: string;
  senderName: string;
  senderRole: CreateMessageBodySenderRole;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface TypeCount {
  projectType: string;
  count: number;
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  completedProjects: number;
  pendingReview: number;
  projectsByStatus: StatusCount[];
  projectsByType: TypeCount[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  project_created: "project_created",
  status_changed: "status_changed",
  file_uploaded: "file_uploaded",
  message_sent: "message_sent",
} as const;

export type ActivityItemActorRole =
  (typeof ActivityItemActorRole)[keyof typeof ActivityItemActorRole];

export const ActivityItemActorRole = {
  client: "client",
  designer: "designer",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  projectId: number;
  projectTitle: string;
  description: string;
  actorName: string;
  actorRole: ActivityItemActorRole;
  createdAt: string;
}

export type GetProjectsParams = {
  status?: GetProjectsStatus;
};

export type GetProjectsStatus =
  (typeof GetProjectsStatus)[keyof typeof GetProjectsStatus];

export const GetProjectsStatus = {
  pending: "pending",
  in_progress: "in_progress",
  review: "review",
  delivered: "delivered",
  completed: "completed",
} as const;