# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the freelance design client portal
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects
 */
export const GetProjectsQueryParams = zod.object({
  status: zod
    .enum(["pending", "in_progress", "review", "delivered", "completed"])
    .optional(),
});

export const GetProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  projectType: zod.enum([
    "logo",
    "branding",
    "web_design",
    "social_media",
    "print",
    "illustration",
    "other",
  ]),
  status: zod.enum([
    "pending",
    "in_progress",
    "review",
    "delivered",
    "completed",
  ]),
  priority: zod.enum(["low", "medium", "high", "urgent"]),
  clientName: zod.string(),
  clientEmail: zod.string(),
  budget: zod.number().nullable(),
  deadline: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  fileCount: zod.number(),
  messageCount: zod.number(),
});
export const GetProjectsResponse = zod.array(GetProjectsResponseItem);

/**
 * @summary Create a new design project request
 */
export const CreateProjectBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  projectType: zod.enum([
    "logo",
    "branding",
    "web_design",
    "social_media",
    "print",
    "illustration",
    "other",
  ]),
  priority: zod.enum(["low", "medium", "high", "urgent"]),
  clientName: zod.string(),
  clientEmail: zod.string(),
  budget: zod.number().nullish(),
  deadline: zod.string().nullish(),
});

/**
 * @summary Get a single project
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod
  .object({
    id: zod.number(),
    title: zod.string(),
    description: zod.string(),
    projectType: zod.enum([
      "logo",
      "branding",
      "web_design",
      "social_media",
      "print",
      "illustration",
      "other",
    ]),
    status: zod.enum([
      "pending",
      "in_progress",
      "review",
      "delivered",
      "completed",
    ]),
    priority: zod.enum(["low", "medium", "high", "urgent"]),
    clientName: zod.string(),
    clientEmail: zod.string(),
    budget: zod.number().nullable(),
    deadline: zod.string().nullable(),
    createdAt: zod.string(),
    updatedAt: zod.string(),
    fileCount: zod.number(),
    messageCount: zod.number(),
  })
  .and(
    zod.object({
      files: zod.array(
        zod.object({
          id: zod.number(),
          projectId: zod.number(),
          fileName: zod.string(),
          fileSize: zod.number(),
          fileType: zod.string(),
          uploadedBy: zod.enum(["client", "designer"]),
          uploadedAt: zod.string(),
          url: zod.string(),
        }),
      ),
      messages: zod.array(
        zod.object({
          id: zod.number(),
          projectId: zod.number(),
          content: zod.string(),
          senderName: zod.string(),
          senderRole: zod.enum(["client", "designer"]),
          createdAt: zod.string(),
        }),
      ),
    }),
  );

/**
 * @summary Update a project (status, notes, etc)
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  status: zod
    .enum(["pending", "in_progress", "review", "delivered", "completed"])
    .optional(),
  priority: zod.enum(["low", "medium", "high", "urgent"]).optional(),
  budget: zod.number().nullish(),
  deadline: zod.string().nullish(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  projectType: zod.enum([
    "logo",
    "branding",
    "web_design",
    "social_media",
    "print",
    "illustration",
    "other",
  ]),
  status: zod.enum([
    "pending",
    "in_progress",
    "review",
    "delivered",
    "completed",
  ]),
  priority: zod.enum(["low", "medium", "high", "urgent"]),
  clientName: zod.string(),
  clientEmail: zod.string(),
  budget: zod.number().nullable(),
  deadline: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  fileCount: zod.number(),
  messageCount: zod.number(),
});

/**
 * @summary Get files for a project
 */
export const GetProjectFilesParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectFilesResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  fileName: zod.string(),
  fileSize: zod.number(),
  fileType: zod.string(),
  uploadedBy: zod.enum(["client", "designer"]),
  uploadedAt: zod.string(),
  url: zod.string(),
});
export const GetProjectFilesResponse = zod.array(GetProjectFilesResponseItem);

/**
 * @summary Upload a file reference for a project
 */
export const UploadProjectFileParams = zod.object({
  id: zod.coerce.number(),
});

export const UploadProjectFileBody = zod.object({
  fileName: zod.string(),
  fileSize: zod.number(),
  fileType: zod.string(),
  uploadedBy: zod.enum(["client", "designer"]),
  url: zod.string(),
});

/**
 * @summary Get messages for a project
 */
export const GetProjectMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectMessagesResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  content: zod.string(),
  senderName: zod.string(),
  senderRole: zod.enum(["client", "designer"]),
  createdAt: zod.string(),
});
export const GetProjectMessagesResponse = zod.array(
  GetProjectMessagesResponseItem,
);

/**
 * @summary Post a message on a project
 */
export const CreateProjectMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateProjectMessageBody = zod.object({
  content: zod.string(),
  senderName: zod.string(),
  senderRole: zod.enum(["client", "designer"]),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalProjects: zod.number(),
  activeProjects: zod.number(),
  completedProjects: zod.number(),
  pendingReview: zod.number(),
  projectsByStatus: zod.array(
    zod.object({
      status: zod.string(),
      count: zod.number(),
    }),
  ),
  projectsByType: zod.array(
    zod.object({
      projectType: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get recent activity feed
 */
export const GetDashboardActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum([
    "project_created",
    "status_changed",
    "file_uploaded",
    "message_sent",
  ]),
  projectId: zod.number(),
  projectTitle: zod.string(),
  description: zod.string(),
  actorName: zod.string(),
  actorRole: zod.enum(["client", "designer"]),
  createdAt: zod.string(),
});
export const GetDashboardActivityResponse = zod.array(
  GetDashboardActivityResponseItem,
);