# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the freelance design client portal
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemActorRole } from "./activityItemActorRole";
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  projectId: number;
  projectTitle: string;
  description: string;
  actorName: string;
  actorRole: ActivityItemActorRole;
  createdAt: string;
}